Refactor GetUnsendNotification Repository

+++ CallbackService/Repository/IDatabaseService.cs
@@ -5,7 +5,7 @@ namespace CallbackService.Repository;
 
 public interface IDatabaseService
 {
-    public long InsertNotification(int cardId, string message);
+    public int InsertNotification(int cardId, string message);
     public IEnumerable <NotificationJoin> GetUnsendNotification();
     public void UpdateNotificationStatus(int notificationId);
 }
